wal is enabled by default

+++ consensus/cbft/state/state_test.go
@@ -200,6 +200,6 @@ func TestNewViewChanges(t *testing.T) {
 
 	assert.Equal(t, 10, viewState.ViewChangeLen())
 	assert.Equal(t, 10, len(viewState.AllViewChange()))
-	assert.Equal(t, 9, viewState.ViewChangeByIndex(9).ValidatorIndex)
+	assert.Equal(t, uint32(9), viewState.ViewChangeByIndex(9).ValidatorIndex)
 
 }
